feat: PathFinder 도메인 객체 단위 테스트 및 통과 로직 작성
