+++ src/main/java/nextstep/subway/line/domain/PathFinder.java
@@ -1,0 +1,12 @@
+package nextstep.subway.line.domain;
+
+import nextstep.subway.station.entity.Station;
+
+import java.math.BigInteger;
+import java.util.List;
+
+public interface PathFinder {
+    List<Station> getPath();
+
+    BigInteger getWeight();
+}
