+++ src/main/java/nextstep/subway/line/domain/entity/Section.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.entity;
+package nextstep.subway.line.domain.entity;
 
 import lombok.AccessLevel;
 import lombok.Getter;
