+++ src/main/java/nextstep/subway/line/domain/entity/handler/SectionAdditionHandlerMapping.java
@@ -1,8 +1,8 @@
-package nextstep.subway.line.entity.handler;
+package nextstep.subway.line.domain.entity.handler;
 
 import nextstep.subway.common.exception.CreationValidationException;
-import nextstep.subway.line.entity.Section;
-import nextstep.subway.line.entity.Sections;
+import nextstep.subway.line.domain.entity.Section;
+import nextstep.subway.line.domain.vo.Sections;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
