+++ src/main/java/nextstep/subway/line/service/PathService.java
@@ -1,11 +1,11 @@
 package nextstep.subway.line.service;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.subway.line.algorithm.ShortestPathFinder;
+import nextstep.subway.line.domain.PathFinder;
+import nextstep.subway.line.domain.ShortestPathFinder;
 import nextstep.subway.line.dto.ShortestPathResponse;
-import nextstep.subway.line.entity.Line;
-import nextstep.subway.line.entity.LineRepository;
-import nextstep.subway.line.entity.Section;
+import nextstep.subway.line.domain.entity.Line;
+import nextstep.subway.line.domain.entity.LineRepository;
 import nextstep.subway.line.exception.StationNotFoundException;
 import nextstep.subway.station.dto.StationResponse;
 import nextstep.subway.station.entity.Station;
@@ -13,7 +13,6 @@ import nextstep.subway.station.entity.StationRepository;
 import org.springframework.stereotype.Service;
 
 import java.math.BigInteger;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,9 +30,9 @@ public class PathService {
 
         List<Line> lineList = lineRepository.findAll();
 
-        ShortestPathFinder pathFinder = new ShortestPathFinder();
-        List<Station> searchedPath = pathFinder.searchPath(lineList, sourceStation, targetStation);
-        BigInteger weight = pathFinder.getShortestWeight(lineList, sourceStation, targetStation);
+        PathFinder pathFinder = new ShortestPathFinder(lineList, sourceStation, targetStation);
+        List<Station> searchedPath = pathFinder.getPath();
+        BigInteger weight = pathFinder.getWeight();
         return new ShortestPathResponse(searchedPath.stream().map(StationResponse::from).collect(Collectors.toList()), weight);
     }
