+++ src/test/java/nextstep/subway/unit/line/handler/SectionAdditionHandlerMappingTest.java
@@ -1,8 +1,8 @@
 package nextstep.subway.unit.line.handler;
 
-import nextstep.subway.line.entity.Line;
-import nextstep.subway.line.entity.Section;
-import nextstep.subway.line.entity.handler.*;
+import nextstep.subway.line.domain.entity.Line;
+import nextstep.subway.line.domain.entity.Section;
+import nextstep.subway.line.domain.entity.handler.*;
 import nextstep.subway.station.entity.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
